fix(abi): correct ctype bit layout to match glibc, add mmap fallback, and fix env reentry
ctype_abi: Fix character classification bitmask constants to match
glibc's actual x86_64 layout where property groups span the low and
high bytes non-sequentially (e.g., _ISUPPER=0x0100, _ISLOWER=0x0200,
_ISBLANK=0x0001). The previous sequential layout (1<<0 through 1<<11)
produced incorrect classification results for the __ctype_b_loc table.

malloc_abi: When the static bump heap is exhausted, fall back to
mmap(MAP_PRIVATE|MAP_ANONYMOUS) via raw syscall instead of returning
NULL. This prevents early startup OOM when the bump arena is consumed
before the membrane pipeline is initialized.

stdlib_abi: Add reentry guards to native_setenv and native_unsetenv
using AtomicBool CAS. Our setenv/unsetenv exports shadow the host's
versioned glibc symbols, so native_*_sym can resolve back to our
own functions causing infinite recursion. On reentry, setenv silently
succeeds and unsetenv falls back to remove_from_environ which directly
manipulates the environ array.

wchar_abi: Remove unused OsStrExt import, fix SYS_openat/SYS_readlinkat
casts to use implicit type instead of explicit `as i64`.

+++ crates/frankenlibc-abi/src/wchar_abi.rs
@@ -4,7 +4,6 @@
 //! On Linux/glibc, `wchar_t` is 32-bit (UTF-32).
 //!
 use std::ffi::{c_char, c_int, c_long, c_longlong, c_ulong, c_ulonglong, c_void};
-use std::os::unix::ffi::OsStrExt;
 
 use frankenlibc_core::stdio::printf::{FormatSegment, Precision, Width, parse_format_string};
 use frankenlibc_membrane::heal::{HealingAction, global_healing_policy};
@@ -1866,7 +1865,7 @@ pub unsafe extern "C" fn realpath(
     // Open the path with O_PATH (no actual I/O, just get an fd for the kernel path).
     let fd = unsafe {
         libc::syscall(
-            libc::SYS_openat as i64,
+            libc::SYS_openat,
             libc::AT_FDCWD,
             path_cstr.as_ptr(),
             libc::O_PATH | libc::O_CLOEXEC,
@@ -1890,14 +1889,14 @@ pub unsafe extern "C" fn realpath(
     let mut buf = [0u8; libc::PATH_MAX as usize];
     let n = unsafe {
         libc::syscall(
-            libc::SYS_readlinkat as i64,
+            libc::SYS_readlinkat,
             libc::AT_FDCWD,
             proc_path.as_ptr(),
             buf.as_mut_ptr(),
             buf.len() - 1,
         ) as isize
     };
-    unsafe { libc::syscall(libc::SYS_close as i64, fd) };
+    unsafe { libc::syscall(libc::SYS_close, fd) };
 
     if n <= 0 {
         unsafe { set_abi_errno(errno::ENOENT) };
